Cleaner string formatting.
This is a code-style patch that

a) transforms all "{format.var}" placeholders into "{format:var}".
b) Allows both format and variable names to be \w+ rather than just [a-z]+
c) Removes the format specifier from the variable name as it is passed to formatText().
d) Cleans up formatText() a bit.